Rename and move files into appropriate packages
